+++ src/hr/avrbanac/openglplayground/OpenGLPlayground.java
@@ -9,6 +9,7 @@ import hr.avrbanac.openglplayground.renderengine.ModelLoader;
 import hr.avrbanac.openglplayground.renderengine.ModelRenderer;
 import hr.avrbanac.openglplayground.models.RawModel;
 import hr.avrbanac.openglplayground.models.TexturedModel;
+import hr.avrbanac.openglplayground.renderengine.OBJLoader;
 import hr.avrbanac.openglplayground.shaders.StaticShader;
 import hr.avrbanac.openglplayground.textures.ModelTexture;
 import static org.lwjgl.glfw.GLFW.glfwSetErrorCallback;
@@ -19,7 +20,7 @@ import org.lwjgl.glfw.GLFWErrorCallback;
  * Main class with application point of entry.
  * 
  * @author avrbanac
- * @version 1.0.0
+ * @version 1.0.1
  */
 public class OpenGLPlayground implements Runnable {
 
@@ -63,123 +64,21 @@ public class OpenGLPlayground implements Runnable {
         DisplayManager dm = new DisplayManager();
         dm.createDisplay();
         
-        ModelLoader loader = new ModelLoader();
-        StaticShader shader = new StaticShader();
-        ModelRenderer renderer = new ModelRenderer(shader, dm.getWidth(), dm.getHeight());
-        // obsolite now that we are using indices
-//        float[] vertices = { 
-//            -0.5f,  0.5f, 0f,
-//            -0.5f, -0.5f, 0f,
-//             0.5f,  0.5f, 0f,
-//             
-//             0.5f,  0.5f, 0f,
-//            -0.5f, -0.5f, 0f,
-//             0.5f, -0.5f, 0f
-//        };
-        // moving on from quad to something more complicated
-//        float[] vertices = {
-//            -0.5f,  0.5f, 0f, //V0
-//            -0.5f, -0.5f, 0f, //V1
-//             0.5f, -0.5f, 0f, //V2
-//             0.5f,  0.5f, 0f, //V3
-//        };
-//        
-//        int[] indices = {
-//            0, 1, 3,
-//            3, 1, 2
-//        };
-//        
-//        float[] textureCoords = {
-//            0,0, //V0
-//            0,1, //V1
-//            1,1, //V2
-//            1,0  //V3
-//        };
-
-        float[] vertices = {			
-            -0.5f,0.5f,-0.5f,	
-            -0.5f,-0.5f,-0.5f,	
-            0.5f,-0.5f,-0.5f,	
-            0.5f,0.5f,-0.5f,		
-
-            -0.5f,0.5f,0.5f,	
-            -0.5f,-0.5f,0.5f,	
-            0.5f,-0.5f,0.5f,	
-            0.5f,0.5f,0.5f,
-
-            0.5f,0.5f,-0.5f,	
-            0.5f,-0.5f,-0.5f,	
-            0.5f,-0.5f,0.5f,	
-            0.5f,0.5f,0.5f,
-
-            -0.5f,0.5f,-0.5f,	
-            -0.5f,-0.5f,-0.5f,	
-            -0.5f,-0.5f,0.5f,	
-            -0.5f,0.5f,0.5f,
-
-            -0.5f,0.5f,0.5f,
-            -0.5f,0.5f,-0.5f,
-            0.5f,0.5f,-0.5f,
-            0.5f,0.5f,0.5f,
-
-            -0.5f,-0.5f,0.5f,
-            -0.5f,-0.5f,-0.5f,
-            0.5f,-0.5f,-0.5f,
-            0.5f,-0.5f,0.5f
-        };
-		
-        float[] textureCoords = {
-            0,0,
-            0,1,
-            1,1,
-            1,0,			
-            0,0,
-            0,1,
-            1,1,
-            1,0,			
-            0,0,
-            0,1,
-            1,1,
-            1,0,
-            0,0,
-            0,1,
-            1,1,
-            1,0,
-            0,0,
-            0,1,
-            1,1,
-            1,0,
-            0,0,
-            0,1,
-            1,1,
-            1,0
-        };
-		
-        int[] indices = {
-            0,1,3,	
-            3,1,2,	
-            4,5,7,
-            7,5,6,
-            8,9,11,
-            11,9,10,
-            12,13,15,
-            15,13,14,	
-            16,17,19,
-            19,17,18,
-            20,21,23,
-            23,21,22
-
-        };
+        ModelLoader loader      = new ModelLoader();
+        StaticShader shader     = new StaticShader();
+        ModelRenderer renderer  = new ModelRenderer(shader, dm.getWidth(), dm.getHeight());
         
-        RawModel model = loader.loadToVAO(vertices, textureCoords, indices);
-        ModelTexture texture = new ModelTexture(loader.loadTexture("dice"));
+        RawModel model          = OBJLoader.loadObjModel("Tree", loader);
+        ModelTexture texture    = new ModelTexture(loader.loadTexture("metal"));
+        
         TexturedModel texturedModel = new TexturedModel(model, texture);
-        Entity entity = new Entity(texturedModel, new Vector3f(0,0,-5), 0, 0, 0, 1);
+        
+        Entity entity = new Entity(texturedModel, new Vector3f(0,0,-50), 0, 0, 0, 1);
         Camera camera = new Camera(dm.getWindow());
         
         while(isRunning) {
             //entity.increasePosition(0, 0, -0.1f);
-            entity.increaseRotation(1, 1, 0);
+            entity.increaseRotation(0, 1, 0);
             camera.move();
             
             renderer.prepare();